requirements.txt
streamlit>=1.35.0
azure-identity>=1.14.0
msgraph-core>=1.0.0

pandas
numpy
requests
openpyxl
python-dotenv
nest_asyncio